spirt

Enum DiagMsgPart

pub enum DiagMsgPart {
    Plain(Cow<'static, str>),
    Attrs(AttrSet),
    Type(Type),
    Const(Const),
    QPtrUsage(QPtrUsage),
}

One part of a Diagnostic message, allowing rich interpolation.

Note: visit::Visitor and transform::Transformer do not interact with any interpolated information, and itโs instead treated as โfrozenโ data.

Variants

Plain(Cow<'static, str>)

Attrs(AttrSet)

Type(Type)

Const(Const)

QPtrUsage(QPtrUsage)