Module spirt::print

Structs

AttrsAndDef

A Print Output type that splits the attributes from the main body of the definition, allowing additional processing before they get concatenated.

ExpectedVsFound

Helper for printing a mismatch error between two nodes (e.g. types), while taking advantage of the print infrastructure that will print all dependencies.

Plan

“Definitions-before-uses” / “topo-sorted” printing plan.

Printer

Enums

AttrStyle

Versions

Wrapper for handling the difference between single-version and multi-version output, which aren’t expressible in pretty::Fragment.

Traits

Print