Struct spirt::print::ExpectedVsFound

pub struct ExpectedVsFound<E, F> {
    pub expected: E,
    pub found: F,
}

Helper for printing a mismatch error between two nodes (e.g. types), while taking advantage of the print infrastructure that will print all dependencies.

Fields
expected: Efound: F