Module spirt::spv::spec::indexed

Utilities for indexing data in a variety of ways (names, compact indices, etc.).

Structs

KhrSegmentedVec

Khronos-oriented segmented sparse array (see KhrSegmented).

NamedIdxMap

Enums

Flat

Flat array (Vec) storage, likely used with compact indices.

FlatWithHoles

Like Flat, but the Vec elements are wrapped in Option.

KhrSegmented

Segmented sparse storage, taking advantage of Khronos’ predictable reservation policy for SPIR-V instruction opcodes and ValueEnums:

Traits

FlatIdx

StorageShape