Struct spirt::spv::spec::Enumerant

pub struct Enumerant {
    pub req_params: ArrayVec<PackedOperandNameAndKind, 3>,
    pub rest_params: Option<OperandKind>,
}

Fields
req_params: ArrayVec<PackedOperandNameAndKind, 3>rest_params: Option<OperandKind>
Implementations

impl Enumerant

pub fn all_params( &self ) -> impl Iterator<Item = (OperandMode, OperandKind)> + '_

Return a (potentially infinite) iterator of OperandKinds, along with the OperandMode indicating whether an operand is expected (Required), or that an operand’s absence signals the end of operands (Optional), which is also the exit signal for the “rest operands” infinite iterators.

pub fn all_params_with_names( &self ) -> impl Iterator<Item = (OperandMode, PackedOperandNameAndKind)> + '_

Like all_params, but providing access to the operand names as well.