Struct spirt::spv::spec::Opcode

pub struct Opcode(/* private fields */);

Implementations

impl Opcode

pub fn try_from_u16_with_name_and_def( opcode: u16 ) -> Option<(Self, &'static str, &'static InstructionDef)>

Lookup the name & definition for opcode in the lazily-loaded Spec, returning None if it’s not a known opcode.

pub fn as_u16(self) -> u16

pub fn name_and_def(self) -> (&'static str, &'static InstructionDef)

Lookup the name & definition for this opcode in the lazily-loaded Spec.

pub fn name(self) -> &'static str

Lookup the name for this opcode in the lazily-loaded Spec.

pub fn def(self) -> &'static InstructionDef

Lookup the definition for this opcode in the lazily-loaded Spec.