Struct spirt::spv::spec::OperandKind

pub struct OperandKind(_);

Implementations

impl OperandKind

pub fn name_and_def(self) -> (&'static str, &'static OperandKindDef)

Lookup the name & definition for this operand kind in the lazily-loaded Spec.

pub fn name(self) -> &'static str

Lookup the name for this operand kind in the lazily-loaded Spec.

pub fn def(self) -> &'static OperandKindDef

Lookup the definition for this operand kind in the lazily-loaded Spec.