Struct spirt::spv::spec::Spec

pub struct Spec {
    pub magic: u32,
    pub well_known: WellKnown,
    pub instructions: NamedIdxMap<Opcode, InstructionDef, KhrSegmented>,
    pub operand_kinds: NamedIdxMap<OperandKind, OperandKindDef, Flat>,
    /* private fields */
}

Fields
magic: u32well_known: WellKnown

Pre-cached IDs for “well-known” names.

instructions: NamedIdxMap<Opcode, InstructionDef, KhrSegmented>operand_kinds: NamedIdxMap<OperandKind, OperandKindDef, Flat>
Implementations

impl Spec

pub fn get() -> &'static Spec

Return a lazily-loaded Spec (only does significant work for the first call).

pub fn get_ext_inst_set_by_lowercase_name( &self, lowercase_ext_inst_set_name: &str ) -> Option<&ExtInstSetDesc>

Return a lazily-parsed ExtInstSetDesc, if a known one exists for this OpExtInstImport name (required to be lowercase, due to Khronos’ choice of case insensitivity, but not checked by this function).