Trait spirt::visit::DynVisit

pub trait DynVisit<'a, V> {
    fn dyn_visit_with(&'a self, visitor: &mut V);
}

Dynamic dispatch version of Visit.

dyn DynVisit<'a, V> is possible, unlike dyn Visit, because of the trait-level type parameter V, which replaces the method parameter.

Required Methods

fn dyn_visit_with(&'a self, visitor: &mut V)