spirt::spv

Module print

Pretty-printing SPIR-V operands.

Structs

TokensForOperand

All the Tokens outputted by printing one single (“logical”) SPIR-V operand, which may be concatenated (after separately processing IDs) to obtain a complete plain-text version of the printed operand.

Enums

Token

The smallest unit produced by printing a (“logical”) SPIR-V operand.

Functions

inst_operands

Group (ordered according to opcode) imms and ids into logical operands (i.e. long immediates are unflattened) and produce one TokensForOperand by printing each of them.

operand_from_imms

Print a single SPIR-V operand from only immediates, potentially composed of an enumerand with parameters (which consumes more immediates).