Module spirt::spv::spec

SPIR-V specification parsing/indexing.

Modules

indexed

Utilities for indexing data in a variety of ways (names, compact indices, etc.).

raw

Deserialization for the .grammar.json files, without any post-processing.

Structs

BitIdx

Enumerant

ExtInstSetDesc

Simplified information for pretty-printing “extended instruction” sets.

ExtInstSetInstructionDesc

Simplified InstructionDef for pretty-printing “extended instruction” sets.

InstructionDef

Opcode

OperandKind

PackedOperandNameAndKind

Spec

WellKnown

Enums

InstructionCategory

LiteralSize

OperandKindDef

OperandMode

RestOperandsUnit

Whether the trailing * “operand” (i.e. repeated arbitrarily many times), consists of just one operand, or two per repeat (used by e.g. OpPhi).

Constants

HEADER_LEN