Struct spirt::ControlRegionDef

pub struct ControlRegionDef {
    pub inputs: SmallVec<[ControlRegionInputDecl; 2]>,
    pub children: EntityList<ControlNode>,
    pub outputs: SmallVec<[Value; 2]>,
}

Definition for a ControlRegion: a control-flow region.

Fields
inputs: SmallVec<[ControlRegionInputDecl; 2]>

Inputs to this ControlRegion:

accessed using Value::ControlRegionInput

values provided by the parent:

when this is the function body: the function’s parameters

children: EntityList<ControlNode>outputs: SmallVec<[Value; 2]>

Output values from this ControlRegion, provided to the parent:

when this is the function body: these are the structured return values

when this is a Select case: these are the values for the parent ControlNode’s outputs (accessed using Value::ControlNodeOutput)

when this is a Loop body: these are the values to be used for the next loop iteration’s body inputs

not accessible through Value::ControlNodeOutput on the Loop, as it’s both confusing regarding Value::ControlRegionInput, and also there’s nothing stopping body-defined values from directly being used outside the loop (once that changes, this aspect can be flipped)