spirt::visit

Trait Visit

pub trait Visit {
    // Required method
    fn visit_with<'a>(&'a self, visitor: &mut impl Visitor<'a>);
}

Trait implemented on “visitable” types (shallowly visitable, at least).

That is, an impl Visit for X will call the relevant Visitor method for X, typically named Visitor::visit_X or Visitor::visit_X_use.

Required Methods

fn visit_with<'a>(&'a self, visitor: &mut impl Visitor<'a>)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.